Implement computer strategy to guess secret color
In previous version, when player is the creator of the secret code,
computer will guess the secret colors by a random color set.

In this change, computer will eliminate color set options from the guess
pool according to feedback of each guess. Thus, it improves the accuracy
of computer's guess.